Rename module file names which don't follow PEP8
