+++ pptx_template/cli.py
@@ -17,7 +17,7 @@ from six import iteritems
 from itertools import islice
 
 from pptx_template.core import edit_slide, remove_slide, get_slide, remove_slide_id, remove_all_slides_having_id
-from pptx_template.xlsxMode import generate_whole_model
+from pptx_template.xlsx_model import generate_whole_model
 
 def process_one_slide(ppt, slide, model):
   if model == u"remove":
