+++ test/test_cli.py
