+++ test/test_text.py
